requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
black==25.1.0
certifi==2025.8.3
click==8.2.1
colorama==0.4.6
Django==5.2.4
django-debug-toolbar==6.0.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-spectacular==0.28.0
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
multidict==6.6.3
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
propcache==0.3.2
PyJWT==2.10.1
python-dotenv==1.1.1
python-telegram-bot==22.3
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.26.0
sniffio==1.3.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
yarl==1.20.1
